+++ plugins/data/dataframes/src/hopeit/dataframes/serialization/dataset.py
@@ -2,7 +2,7 @@
 """
 
 from importlib import import_module
-from typing import Type, TypeVar
+from typing import Generic, Type, TypeVar
 
 from hopeit.dataobjects import dataclass, dataobject
 
@@ -11,11 +11,19 @@ DataFrameT = TypeVar("DataFrameT")
 
 @dataobject
 @dataclass
-class Dataset:
+class Dataset(Generic[DataFrameT]):
+    """Persisted representation of a @dataframe object"""
     protocol: str
     partition_key: str
     key: str
     datatype: str
+
+    async def load(self) -> DataFrameT:
+        return await self.__storage.load(self)  # type: ignore[attr-defined]
+
+    @classmethod
+    async def save(cls, dataframe: DataFrameT) -> "Dataset[DataFrameT]":
+        return await cls.__storage.save(dataframe)  # type: ignore[attr-defined]
 
 
 def find_protocol_impl(qual_type_name: str) -> Type: